backend/innovo_backend/shared/jwt_utils.py
"""JWT token utilities for authentication."""
import os
from datetime import datetime, timedelta
from typing import Optional, Dict
import jwt
from jwt.exceptions import InvalidTokenError, ExpiredSignatureError
import logging

logger = logging.getLogger(__name__)

SECRET_KEY = os.getenv("JWT_SECRET_KEY")
if not SECRET_KEY:
    raise ValueError(
        "JWT_SECRET_KEY environment variable is required. "
        "Generate one using: openssl rand -hex 32"
    )
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_HOURS = 24


def create_access_token(data: Dict[str, str]) -> str:
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(hours=ACCESS_TOKEN_EXPIRE_HOURS)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    logger.info(f"Access token created for user: {data.get('email', 'unknown')}")
    return encoded_jwt


def verify_token(token: str) -> Optional[Dict[str, str]]:
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        return payload
    except ExpiredSignatureError:
        logger.warning("Attempted to use expired token")
        return None
    except InvalidTokenError as e:
        logger.warning(f"Invalid token provided: {str(e)}")
        return None
    except Exception as e:
        logger.error(f"Error verifying token: {str(e)}")
        return None


def create_password_reset_token(email: str) -> str:
    expire = datetime.utcnow() + timedelta(hours=1)
    to_encode = {
        "email": email,
        "type": "password_reset",
        "exp": expire,
    }
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    logger.info(f"Password reset token created for: {email}")
    return encoded_jwt


def verify_password_reset_token(token: str) -> Optional[str]:
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        if payload.get("type") != "password_reset":
            logger.warning("Token provided is not a password reset token")
            return None
        email = payload.get("email")
        if not email:
            logger.warning("Password reset token missing email")
            return None
        return email
    except ExpiredSignatureError:
        logger.warning("Password reset token has expired")
        return None
    except InvalidTokenError as e:
        logger.warning(f"Invalid password reset token: {str(e)}")
        return None
    except Exception as e:
        logger.error(f"Error verifying password reset token: {str(e)}")
        return None
