backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "innovo_backend"
version = "1.0.0"
description = "Innovo funding-document AI backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "sqlalchemy==2.0.36",
    "alembic>=1.13.0",
    "psycopg2-binary==2.9.9",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.22",
    "python-dotenv==1.0.1",
    "passlib[bcrypt]==1.7.4",
    "pyjwt>=2.9.0",
    "python-jose[cryptography]>=3.4.0",
    "openai==1.54.0",
    "httpx==0.27.2",
    "supabase>=2.27.2",
    "requests>=2.32.4",
    "beautifulsoup4==4.12.3",
    "reportlab==4.0.7",
    "python-docx==1.1.0",
    "PyPDF2>=3.0.0,<4.0.0",
    "pgvector>=0.3.0",
    "apscheduler>=3.10.0",
    "posthog>=3.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["innovo_backend*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
ignore = ["B008", "B904"]
